requirements.txt
# Database Drivers
psycopg2-binary==2.9.9
neo4j==5.15.0
gremlinpython==3.7.1

# Data Processing
pandas==2.1.4
polars==0.20.2
pyarrow==14.0.2

# SQL Parsing & Schema Analysis
sqlalchemy==2.0.25
sqlparse==0.4.4
sqlglot==20.11.0

# LLM & Embeddings
openai>=1.12.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.12
langgraph==0.0.20
tiktoken==0.5.2

# Vector & Embeddings
sentence-transformers==2.2.2
numpy==1.26.3

# API & Web
fastapi==0.108.0
uvicorn==0.25.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
pyyaml==6.0.1
tenacity==8.2.3

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.1
ruff==0.1.9
mypy==1.8.0

# Jupyter for examples
jupyter==1.0.0
ipykernel==6.28.0
